requirements.txt
pygame>=1.9.1
pillow
psutil
rpi_blacklight
RPi.GPIO
python-vlc
